backend/pyproject.toml
[project]
name = "maas-ultimate-bug-triager"
version = "0.1.0"
description = "Triage MAAS bugs, reproduce them, and suggest fixes"
readme = "README.md"
authors = [
    { name = "Bjorn Tillenius", email = "bjorn@canonical.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "httpx>=0.28",
    "launchpadlib>=1.11",
    "google-genai>=1.0",
    "pydantic>=2.11",
    "pyyaml>=6.0",
]

[project.scripts]
maas-triager = "maas_ultimate_bug_triager.main:main"

[dependency-groups]
dev = [
    "pytest>=8.4",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
]

[build-system]
requires = ["uv_build>=0.11.13,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
